优化日志，新增常量类，新增Task

+++ Distributed Systems/Util/IO/FileUtil.cs
@@ -27,17 +27,19 @@ namespace Util.IO
         /// </summary>
         /// <param name="oldFilePath">源文件</param>
         /// <param name="newFilePath">目标文件</param>
-        public static void Copy(string oldFilePath, string newFilePath)
+        public static bool Copy(string oldFilePath, string newFilePath)
         {
             lock (locker)
             {
-                if (!Exists(oldFilePath))
-                    throw new Util.Exception.ExceptionUtil("源文件不存在！");
-                if (Exists(newFilePath))
-                    throw new Util.Exception.ExceptionUtil("新文件已存在！");
+                if (!Exists(oldFilePath) || Exists(newFilePath))
+                {
+                    return false;
+                }
 
                 SIO.File.Copy(oldFilePath, newFilePath);
             }
+
+            return true;
         }
 
         /// <summary>
@@ -105,17 +107,19 @@ namespace Util.IO
         /// </summary>
         /// <param name="oldFilePath">源文件</param>
         /// <param name="newFilePath">目标文件</param>
-        public static void Move(string oldFilePath, string newFilePath)
+        public static bool Move(string oldFilePath, string newFilePath)
         {
-            if (!Exists(oldFilePath))
-                throw new Util.Exception.ExceptionUtil("源文件不存在！");
-            if (Exists(newFilePath))
-                throw new Util.Exception.ExceptionUtil("新文件已存在！");
+            if (!Exists(oldFilePath) || Exists(newFilePath))
+            {
+                return false;
+            }
 
             lock (locker)
             {
                 SIO.File.Move(oldFilePath, newFilePath);
             }
+
+            return true;
         }
 
         /// <summary>
@@ -142,9 +146,9 @@ namespace Util.IO
         private static void SaveFile(string filePath, string content, Encoding encoding, bool isOverride = false, bool isNewLine = true)
         {
             if (string.IsNullOrEmpty(filePath) || string.IsNullOrEmpty(filePath))
-                throw new Util.Exception.ExceptionUtil("操作不正确！");
-            string dirPath = DirectoryUtil.GetDirPath(filePath);
-            DirectoryUtil.Create(dirPath);
+                throw new System.Exception("操作不正确！");
+            string dirPath = GetDirPath(filePath);
+            SIO.Directory.CreateDirectory(dirPath);
 
             var fileModel = SIO.FileMode.OpenOrCreate;
             if (isOverride)
@@ -153,6 +157,9 @@ namespace Util.IO
             }
             lock (locker)
             {
+                if (!Exists(filePath))
+                    Create(filePath);
+
                 using (SIO.FileStream file = new SIO.FileStream(filePath, fileModel))
                 {
                     encoding = encoding == null ? Encoding.UTF8 : encoding;
@@ -180,6 +187,19 @@ namespace Util.IO
         #endregion
 
         #region Select
+
+        /// <summary>
+        /// 获取路径中的目录部分
+        /// </summary>
+        /// <param name="filePath">文件路径</param>
+        /// <returns></returns>
+        public static string GetDirPath(string filePath)
+        {
+            string path = string.Empty;
+            if (!string.IsNullOrEmpty(filePath))
+                path = filePath.Substring(0, filePath.LastIndexOf(@"\"));
+            return path;
+        }
 
         /// <summary>
         /// 读取文件内容
@@ -291,7 +311,7 @@ namespace Util.IO
         public static string GetFileName(string filePath, bool isContansExtension = true)
         {
             string fileName = string.Empty;
-            
+
             fileName = SIO.Path.GetFileName(filePath);
             if (!isContansExtension)
             {
